Rethinking Html rendering (SSR)
