--- tyrian/jvm/src/test/scala/tyrian/PlaceholderTests.scala
@@ -1,9 +1,0 @@
-package tyrian
-
-class PlaceholderTests extends munit.FunSuite {
-
-  test("placeholder jvm") {
-    assert(1 == 1)
-  }
-
-}
